fix: 수입/지출 금액 타입을 int(원)로 변경해 부동소수점 반올림 오류 해결

+++ lib/features/expense/data/models/expense_model.dart
@@ -16,7 +16,7 @@ sealed class ExpenseModel with _$ExpenseModel {
     String? userId,
     String? accountBookId,
     String? fundingSource,
-    required double amount,
+    required int amount,
     required DateTime date,
     String? category,
     String? merchant,
